frontend: NOT BUILDABLE introduce buffer list and window to reflect ui

+++ yeet-buffer/src/update/viewport.rs
@@ -10,8 +10,8 @@ pub fn update_by_cursor(viewport: &mut ViewPort, cursor: &Cursor, buffer: &TextB
 
     if viewport.vertical_index > cursor.vertical_index {
         viewport.vertical_index = cursor.vertical_index;
-    } else if viewport.vertical_index + (viewport.height - 1) < cursor.vertical_index {
-        viewport.vertical_index = cursor.vertical_index - (viewport.height - 1);
+    } else if viewport.vertical_index + usize::from(viewport.height - 1) < cursor.vertical_index {
+        viewport.vertical_index = cursor.vertical_index - usize::from(viewport.height - 1);
     }
 
     let cursor_index = match cursor.horizontal_index {
@@ -51,10 +51,10 @@ pub fn update_by_direction(
     match direction {
         ViewPortDirection::BottomOnCursor => {
             if let Some(cursor) = &cursor {
-                if cursor.vertical_index < viewport.height {
+                if cursor.vertical_index < usize::from(viewport.height) {
                     viewport.vertical_index = 0;
                 } else {
-                    let index = cursor.vertical_index - viewport.height + 1;
+                    let index = cursor.vertical_index - usize::from(viewport.height) + 1;
                     viewport.vertical_index = index;
                 }
             }
@@ -62,47 +62,47 @@ pub fn update_by_direction(
         ViewPortDirection::CenterOnCursor => {
             if let Some(cursor) = &cursor {
                 let index_offset = viewport.height / 2;
-                if cursor.vertical_index < index_offset {
+                if cursor.vertical_index < usize::from(index_offset) {
                     viewport.vertical_index = 0;
                 } else {
-                    viewport.vertical_index = cursor.vertical_index - index_offset;
+                    viewport.vertical_index = cursor.vertical_index - usize::from(index_offset);
                 }
             }
         }
         ViewPortDirection::HalfPageDown => {
             let index_offset = viewport.height / 2;
-            let viewport_end_index = viewport.vertical_index + (viewport.height - 1);
-            let viewport_end_after_move_index = viewport_end_index + index_offset;
+            let viewport_end_index = viewport.vertical_index + usize::from(viewport.height - 1);
+            let viewport_end_after_move_index = viewport_end_index + usize::from(index_offset);
 
             if viewport_end_after_move_index < buffer.lines.len() {
-                viewport.vertical_index += index_offset;
-            } else if viewport.height > buffer.lines.len() {
+                viewport.vertical_index += usize::from(index_offset);
+            } else if usize::from(viewport.height) > buffer.lines.len() {
                 viewport.vertical_index = 0;
             } else {
-                viewport.vertical_index = buffer.lines.len() - viewport.height;
+                viewport.vertical_index = buffer.lines.len() - usize::from(viewport.height);
             }
 
             if let Some(ref mut cursor) = cursor {
-                if cursor.vertical_index + index_offset >= buffer.lines.len() {
+                if cursor.vertical_index + usize::from(index_offset) >= buffer.lines.len() {
                     cursor.vertical_index = buffer.lines.len() - 1;
                 } else {
-                    cursor.vertical_index += index_offset;
+                    cursor.vertical_index += usize::from(index_offset);
                 }
             }
         }
         ViewPortDirection::HalfPageUp => {
             let index_offset = viewport.height / 2;
-            if viewport.vertical_index < index_offset {
+            if viewport.vertical_index < usize::from(index_offset) {
                 viewport.vertical_index = 0;
             } else {
-                viewport.vertical_index -= index_offset;
+                viewport.vertical_index -= usize::from(index_offset);
             }
 
             if let Some(ref mut cursor) = cursor {
-                if cursor.vertical_index < index_offset {
+                if cursor.vertical_index < usize::from(index_offset) {
                     cursor.vertical_index = 0;
                 } else {
-                    cursor.vertical_index -= index_offset;
+                    cursor.vertical_index -= usize::from(index_offset);
                 }
             }
         }
